requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
SQLAlchemy==2.0.32
psycopg[binary]==3.2.1
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.8.2
prometheus-client==0.20.0
